requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.1.3
django-filter==22.1
djangorestframework==3.12
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
celery==5.2.6
drf_yasg==1.20.0
PyJWT==2.4.0
redis==4.3.1
django-cors-headers==3.14.0
gunicorn==20.0.4
beautifulsoup4==4.10.0
fuzzywuzzy==0.18.0
requests==2.31.0
bs4